backend/requirements.txt
fastapi>=0.123.0,<0.124.0
uvicorn[standard]>=0.34.0,<0.45.0
websockets>=15.0.1,<16.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
google-genai>=1.0.0
